Add files via upload
This folder contains the cleaned code that was used for the Master thesis version. The required data can be found in the subfolder "data", even shorter code without model selection or alternative models can be found in the subfolder "extremely_cleaned_code" for all files for which shortening was possible.